Extract into WynntilsAccount, rename it to WynntilsAccountManager and move it to a better location
