+++ common/src/main/java/com/wynntils/core/webapi/WebManager.java
@@ -8,12 +8,9 @@ import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.managers.CoreManager;
-import com.wynntils.core.webapi.account.WynntilsAccount;
 import com.wynntils.core.webapi.request.RequestBuilder;
 import com.wynntils.core.webapi.request.RequestHandler;
 import com.wynntils.mc.event.WebSetupEvent;
-import com.wynntils.mc.utils.ComponentUtils;
-import com.wynntils.mc.utils.McUtils;
 import com.wynntils.wynn.model.discoveries.objects.DiscoveryInfo;
 import com.wynntils.wynn.objects.profiles.DiscoveryProfile;
 import java.io.File;
@@ -24,11 +21,6 @@ import java.net.URLConnection;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.ClickEvent;
-import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.network.chat.Style;
-import net.minecraft.network.chat.TextComponent;
 
 /** Provides and loads web content on demand */
 public final class WebManager extends CoreManager {
@@ -45,8 +37,6 @@ public final class WebManager extends CoreManager {
 
     private static List<DiscoveryInfo> discoveryInfoList = new ArrayList<>();
 
-    private static WynntilsAccount account = null;
-
     private static final String USER_AGENT = String.format(
             "Wynntils Artemis\\%s (%s) %s",
             WynntilsMod.getVersion(),
@@ -55,41 +45,11 @@ public final class WebManager extends CoreManager {
 
     public static void init() {
         tryReloadApiUrls(false);
-        setupUserAccount();
-    }
-
-    public static boolean isLoggedIn() {
-        return (account != null && account.isConnected());
     }
 
     public static void reset() {
         // tryReloadApiUrls
         apiUrls = null;
-    }
-
-    private static void setupUserAccount() {
-        if (isLoggedIn()) return;
-
-        account = new WynntilsAccount();
-        boolean accountSetup = account.login();
-
-        if (!accountSetup) {
-            MutableComponent failed = new TextComponent(
-                            "Welps! Trying to connect and set up the Wynntils Account with your data has failed. "
-                                    + "Most notably, cloud config syncing will not work. To try this action again, run ")
-                    .withStyle(ChatFormatting.GREEN);
-            failed.append(new TextComponent("/wynntils reload")
-                    .withStyle(Style.EMPTY
-                            .withColor(ChatFormatting.AQUA)
-                            .withClickEvent(new ClickEvent(ClickEvent.Action.RUN_COMMAND, "/wynntils reload"))));
-
-            if (McUtils.player() == null) {
-                WynntilsMod.error(ComponentUtils.getUnformatted(failed));
-                return;
-            }
-
-            McUtils.sendMessageToClient(failed);
-        }
     }
 
     private static void tryReloadApiUrls(boolean async) {
@@ -143,10 +103,6 @@ public final class WebManager extends CoreManager {
         return apiUrls.get(key);
     }
 
-    public static boolean isAthenaOnline() {
-        return (account != null && account.isConnected());
-    }
-
     public static List<DiscoveryInfo> getDiscoveryInfoList() {
         return discoveryInfoList;
     }
@@ -161,10 +117,6 @@ public final class WebManager extends CoreManager {
 
     public static Optional<WebReader> getApiUrls() {
         return Optional.ofNullable(apiUrls);
-    }
-
-    public static Optional<WynntilsAccount> getAccount() {
-        return Optional.ofNullable(account);
     }
 
     public static RequestHandler getHandler() {
