+++ common/src/main/java/com/wynntils/sockets/HadesClientHandler.java
@@ -5,7 +5,7 @@
 package com.wynntils.sockets;
 
 import com.wynntils.core.WynntilsMod;
-import com.wynntils.core.webapi.WebManager;
+import com.wynntils.core.net.athena.WynntilsAccountManager;
 import com.wynntils.features.user.HadesFeature;
 import com.wynntils.hades.objects.HadesConnection;
 import com.wynntils.hades.protocol.interfaces.adapters.IHadesClientAdapter;
@@ -35,7 +35,7 @@ public class HadesClientHandler implements IHadesClientAdapter {
 
     @Override
     public void onConnect() {
-        if (WebManager.getAccount().isEmpty()) {
+        if (WynntilsAccountManager.getAccount().isEmpty()) {
             hadesConnection.disconnect();
 
             if (WorldStateManager.onServer()) {
@@ -47,8 +47,8 @@ public class HadesClientHandler implements IHadesClientAdapter {
             throw new IllegalStateException("Tried to auth to HadesServer when WebManager#getAccount is empty.");
         }
 
-        hadesConnection.sendPacketAndFlush(
-                new HCPacketAuthenticate(WebManager.getAccount().get().getToken()));
+        hadesConnection.sendPacketAndFlush(new HCPacketAuthenticate(
+                WynntilsAccountManager.getAccount().get().getToken()));
     }
 
     @Override
